Added command-line-args package and refactored code
