+++ package.json
@@ -4,17 +4,17 @@
   "description": "",
   "main": "bundler.js",
   "scripts": {
-    "bundle": "node bundler.js"
+    "bundle": "node bundler.js --entryFile ./src/app.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-
   },
   "devDependencies": {
     "@babel/core": "^7.17.9",
     "chalk": "^4.0.0",
-    "terser": "^5.12.1"
+    "terser": "^5.12.1",
+    "command-line-args": "^5.2.1"
   }
 }
